+++ app/Assets/weatherDisplay.js
